backend/import-products/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3